backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.41
alembic==1.15.2
pgvector==0.3.6
pyjwt==2.10.1
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
google-genai==1.14.0
